jdbc/src/main/java/com/jdbc/service/MemberServiceV1.java
package com.jdbc.service;

import com.jdbc.domain.Member;
import com.jdbc.repository.MemberRepositoryV1;
import lombok.RequiredArgsConstructor;

import java.sql.SQLException;

@RequiredArgsConstructor
public class MemberServiceV1 {
    private final MemberRepositoryV1 repositoryV1;

    public void accountTransfer(String fromId, String toId, int money) throws SQLException {
        Member fromMember = repositoryV1.findById(fromId);
        Member toMember = repositoryV1.findById(toId);

        repositoryV1.update(fromId, fromMember.getMoney() - money);
        validation(toMember);
        repositoryV1.update(toId, toMember.getMoney() + money);
    }

    private static void validation(Member toMember) {
        if (toMember.getMemberId().equals("ex")) {
            throw new IllegalStateException("이체중 예외 발생");
        }
    }
}
